requirements.txt
# Зависимости проекта
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-telegram-bot>=20.0
